Rename forlder after gem name
